rename package name
back -> cosmos
