+++ src/main/java/com/_404/cosmos/BackApplication.java
@@ -1,8 +1,7 @@
-package com._404.back;
+package com._404.cosmos;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @SpringBootApplication
 public class BackApplication {
