+++ src/test/java/com/_404/cosmos/BackApplicationTests.java
@@ -1,4 +1,4 @@
-package com._404.back;
+package com._404.cosmos;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
